+++ src/flows/register/types.ts
@@ -1,0 +1,5 @@
+export interface RegsiterParams {
+  programModAddress?: string
+  // TODO: Export ProgramInstance type from sdk
+  programData?: any
+}